Refactor : code refactor
